@@ -2,7 +2,7 @@
 @inject NavigationManager navManager
 @inject WindowManager windowManager
 
-<div class="col-md-12">
+<div class="col-md-12 modal-bg">
 	<section>
 		<h2>Create a new Planet</h2>
 
