@@ -1,20 +1,24 @@
 ﻿@using System.Net.Http.Headers
 @inject HttpClient http
 
-<h3>Basic Planet Information</h3>
+<h3>Planet Information</h3>
 <hr />
 <br />
 
 <span class="text-danger">@resultSpan</span>
 <br />
 
-<p>Name</p>
-<input class="form-control" type="text" @bind-value="@nameValue" />
+<div class="form-group mt-2 new-input-style">
+    <label>Name</label>
+    <input class="form-control" placeholder="Name" @bind-value="@nameValue" />
+</div>
 
 <br />
 
-<p>Description</p>
-<input class="form-control" type="text" @bind-value="@descValue" />
+<div class="form-group mt-2 new-input-style">
+    <label>Description </label>
+    <input class="form-control" placeholder="Description" @bind-value="@descValue" />
+</div>
 
 <br />
 
