@@ -5,9 +5,9 @@
 	{
 		<EmbedRowComponent Row="@row" Message="@Message" embed="@embed" FormItem="@item"></EmbedRowComponent>
 	}
-	@foreach (EmbedItem item in embed.CurrentlyDisplayed.Items)
+	@foreach (EmbedItem _item in item.Items)
 	{
-		<EmbedItemComponent embed="@embed" Message="@Message" item="@item"></EmbedItemComponent>
+		<EmbedItemComponent embed="@embed" Message="@Message" item="@_item" FormItem="@item"></EmbedItemComponent>
 	}
 </div>
 
