@@ -77,7 +77,7 @@
 
                         @if (MessageData.Message.Embed != null && MessageData.Message.Embed.Pages.Count > 0)
                         {
-                            <EmbedComponent embed="@MessageData.Message.Embed" Message="MessageData"></EmbedComponent>
+                            <EmbedComponent Message="MessageData"></EmbedComponent>
                         }
                         else
                         {
