@@ -7,7 +7,7 @@
     return;
 }
 
-<div class="col-md-12">
+<div class="col-md-12 modal-bg">
 	<section style="text-align: center;">
 		<h2>Create a new App</h2>
 
