@@ -1,4 +1,4 @@
-﻿<div class="col-md-12">
+﻿<div class="col-md-12 modal-bg">
 	<section>
 		<h2>Your Friends</h2>
 
