@@ -1,6 +1,7 @@
 @inject HttpClient Http
 @inject NavigationManager navManager
 @inject ILocalStorageService localStorage
+@inject IJSRuntime JS
 
 <body class="login-background">
     <div class="popup-box">
@@ -86,7 +87,8 @@
             if (result.Success)
             {
                 resultSpan = result.Message;
-                await ClientUserManager.StoreToken(localStorage);
+                await JS.SetCookieAsync("token", tokenResult.Data);
+
                 if (FromVerified)
                     navManager.NavigateTo("/", true);
             }
