@@ -68,12 +68,6 @@
 	{
 		await base.OnInitializedAsync();
 
-		var isMobile = await JS.InvokeAsync<bool>("IsMobile");
-		var isEmbedded = await JS.InvokeAsync<bool>("IsEmbedded");
-
-		DeviceInfo.IsMobile = isMobile;
-		DeviceInfo.IsEmbedded = isEmbedded;
-
 		windowManager.OnPlanetFocused += OnPlanetChange;
 		ValourClient.OnLogin += OnLoginAsync;
 	}
@@ -92,6 +86,12 @@
 	{
 		await JS.InvokeVoidAsync("FixClip");
 		await JS.InvokeVoidAsync("FitMobile");
+
+		var isMobile = await JS.InvokeAsync<bool>("IsMobile");
+		var isEmbedded = await JS.InvokeAsync<bool>("IsEmbedded");
+
+		DeviceInfo.IsMobile = isMobile;
+		DeviceInfo.IsEmbedded = isEmbedded;
 	}
 
 	public async Task OnPlanetChange(Planet planet)
