@@ -4,9 +4,9 @@
     ViewData["Title"] = "Valour";
 }
 <div class='index-back'>
-    <div class="text-center">
+    <div class="text-center mt-4">
         <h1 class="display-4">Valour is coming.</h1>
-        <p style='margin: 0'>Join the movement for a more flexible, transparent, and open platform made by communities for communities.</p>
+        <p class="m-2">Join the movement for a more flexible, transparent, and open platform made by communities for communities.</p>
         <a class='v-link' href="#">What is Valour?</a>
     </div>
 
